Refactor /dataclasses/
-100 coolness :(
