+++ app/build.gradle.kts
@@ -184,7 +184,7 @@ sonar {
                 "**/com/se2gruppe5/risikofrontend/Constants.kt",
                 "**/com/se2gruppe5/risikofrontend/game/dialogues/**",
                 "**/com/se2gruppe5/risikofrontend/game/popup/**",
-                "**/com/se2gruppe5/risikofrontend/game/engine/**",
+                "**/com/se2gruppe5/risikofrontend/game/dataclasses/util/**",
                 "**/com/se2gruppe5/risikofrontend/game/enums/**",
                 "**/com/se2gruppe5/risikofrontend/game/managers/GameViewManager.kt",
                 "**/com/se2gruppe5/risikofrontend/game/managers/ToastUtilAndroid.kt"
