+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/board/BoardLoader.kt
@@ -1,9 +1,9 @@
 package com.se2gruppe5.risikofrontend.game.board
 
-import com.se2gruppe5.risikofrontend.game.dataclasses.TerritoryRecord
-import com.se2gruppe5.risikofrontend.game.engine.Point2D
-import com.se2gruppe5.risikofrontend.game.engine.Size2D
-import com.se2gruppe5.risikofrontend.game.engine.Transform2D
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.TerritoryRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.util.Point2D
+import com.se2gruppe5.risikofrontend.game.dataclasses.util.Size2D
+import com.se2gruppe5.risikofrontend.game.dataclasses.util.Transform2D
 import kotlinx.serialization.json.Json
