+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/board/BoardLoaderAndroid.kt
@@ -1,7 +1,7 @@
 package com.se2gruppe5.risikofrontend.game.board
 
 import android.content.Context
-import com.se2gruppe5.risikofrontend.game.dataclasses.TerritoryRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.TerritoryRecord
 import java.io.BufferedReader
 import java.io.InputStreamReader
