+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/cards/ICardHandler.kt
@@ -1,7 +1,7 @@
 package com.se2gruppe5.risikofrontend.game.cards
 
-import com.se2gruppe5.risikofrontend.game.dataclasses.CardRecord
-import com.se2gruppe5.risikofrontend.game.dataclasses.PlayerRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.CardRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.PlayerRecord
 
 interface ICardHandler {
