+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/dataclasses/game/CardRecord.kt
@@ -1,4 +1,4 @@
-package com.se2gruppe5.risikofrontend.game.dataclasses
+package com.se2gruppe5.risikofrontend.game.dataclasses.game
 
 import com.se2gruppe5.risikofrontend.game.enums.CardType
