+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/dataclasses/util/Point2D.kt
@@ -1,0 +1,4 @@
+package com.se2gruppe5.risikofrontend.game.dataclasses.util
+
+data class Point2D(var x: Float, var y: Float)
+
