+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/dataclasses/util/Size2D.kt
@@ -1,0 +1,3 @@
+package com.se2gruppe5.risikofrontend.game.dataclasses.util
+
+data class Size2D(var x: Float, var y: Float)
