+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/dataclasses/util/Transform2D.kt
@@ -1,4 +1,4 @@
-package com.se2gruppe5.risikofrontend.game.engine
+package com.se2gruppe5.risikofrontend.game.dataclasses.util
 
 data class Transform2D(var position: Point2D, var size: Size2D)
