--- app/src/main/java/com/se2gruppe5/risikofrontend/game/engine/Point2D.kt
@@ -1,4 +1,0 @@
-package com.se2gruppe5.risikofrontend.game.engine
-
-data class Point2D(var x: Float, var y: Float)
-
