--- app/src/main/java/com/se2gruppe5/risikofrontend/game/engine/Size2D.kt
@@ -1,3 +1,0 @@
-package com.se2gruppe5.risikofrontend.game.engine
-
-data class Size2D(var x: Float, var y: Float)
