+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/managers/GameManager.kt
@@ -1,7 +1,7 @@
 package com.se2gruppe5.risikofrontend.game.managers
 
-import com.se2gruppe5.risikofrontend.game.dataclasses.PlayerRecord
-import com.se2gruppe5.risikofrontend.game.dataclasses.TerritoryRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.PlayerRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.TerritoryRecord
 import com.se2gruppe5.risikofrontend.game.enums.Phases
 import com.se2gruppe5.risikofrontend.network.INetworkClient
 import java.util.UUID
