+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/managers/GameViewManager.kt
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import com.se2gruppe5.risikofrontend.R
 import com.se2gruppe5.risikofrontend.game.board.BoardVisualGeneratorAndroid
-import com.se2gruppe5.risikofrontend.game.dataclasses.PlayerRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.PlayerRecord
 import com.se2gruppe5.risikofrontend.game.territory.IPointingArrowUI
 import com.se2gruppe5.risikofrontend.game.territory.PointingArrowAndroid
 import java.util.UUID
