+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/territory/IPointingArrowUI.kt
@@ -1,6 +1,6 @@
 package com.se2gruppe5.risikofrontend.game.territory
 
-import com.se2gruppe5.risikofrontend.game.engine.Point2D
+import com.se2gruppe5.risikofrontend.game.dataclasses.util.Point2D
 
 interface IPointingArrowUI {
     fun setColor(color: Int)
