+++ app/src/main/java/com/se2gruppe5/risikofrontend/game/territory/LineAndroid.kt
@@ -5,7 +5,7 @@ import android.graphics.Canvas
 import android.graphics.Paint
 import android.util.AttributeSet
 import android.view.View
-import com.se2gruppe5.risikofrontend.game.engine.Point2D
+import com.se2gruppe5.risikofrontend.game.dataclasses.util.Point2D
 
 const val LINE_STROKE_COLOR: Int = 0x000000
 const val LINE_STROKE_WIDTH: Float = 5f
