+++ app/src/main/java/com/se2gruppe5/risikofrontend/network/INetworkClient.kt
@@ -1,8 +1,8 @@
 package com.se2gruppe5.risikofrontend.network
 
-import com.se2gruppe5.risikofrontend.game.dataclasses.CardRecord
-import com.se2gruppe5.risikofrontend.game.dataclasses.PlayerRecord
-import com.se2gruppe5.risikofrontend.game.dataclasses.TerritoryRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.CardRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.PlayerRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.TerritoryRecord
 import java.util.UUID
 
 interface INetworkClient {
