+++ app/src/main/java/com/se2gruppe5/risikofrontend/network/sse/messages/GameStartMessage.kt
@@ -1,6 +1,6 @@
 package com.se2gruppe5.risikofrontend.network.sse.messages
 
-import com.se2gruppe5.risikofrontend.game.dataclasses.PlayerRecord
+import com.se2gruppe5.risikofrontend.game.dataclasses.game.PlayerRecord
 import com.se2gruppe5.risikofrontend.network.sse.IMessage
 import java.io.Serializable
 import java.util.UUID
